+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add SOLR_CORES env variable on solr container to define the core(s) name. (default = "collection1")
 - Add DATABASE_PREFIXES env variable on engine container. This var define the prefix used to name the db connection vars (default = "DATABASE"). It's possible to define multiple prefixes to handle multiple db.
+- Remove composer.phar manipulation and add composer in the engine image.
 
 ### 1.3.0